+++ src/main/java/racingcar/accessor/GameAccessor.java
@@ -1,4 +1,4 @@
-package racingcar.service;
+package racingcar.accessor;
 
 import racingcar.domain.Cars;
 import racingcar.dto.CarsDto;
@@ -6,7 +6,7 @@ import racingcar.dto.WinnersDto;
 
 import java.util.List;
 
-public interface GameService {
+public interface GameAccessor {
     void setUpGame(List<String> userNames, int goalRound);
 
     void setUpGame(Cars cars, int goalRound);
